+++ doitmyself_team/src/main/java/com/kh/dim2/DAO/adminDAO.java
@@ -141,11 +141,5 @@ public class adminDAO {
 		return sqlSession.selectList("Admin.getExchangeList");
 	}
 	
-	public List<Product> getProductList(Map<String, Object> map) {
-		return sqlSession.selectList("Admin.productList",map);
-	}
 
-	public int getProductListcount(Map<String, Object> map) {
-		return sqlSession.selectOne("Admin.getProductListcount",map);
-	}
 }
